Refactored BaseNetworkClient based on PR comments
