+++ QuizApp/Client/Login/LoginEndpoints.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+enum LoginEndpoints: String {
+
+    case login
+
+    var path: URL {
+        switch self {
+        case .login:
+            return Api.apiURL.appendingPathComponent(self.rawValue)
+        }
+    }
+
+}
