+++ QuizApp/Client/Login/LoginNetworkClient.swift
@@ -1,35 +1,12 @@
 import Foundation
 
-enum LoginEndpoints: String {
-
-    case login
-
-    var path: URL {
-        switch self {
-        case .login:
-            return Api.apiURL.appendingPathComponent(self.rawValue)
-        }
-    }
-
-}
-
-protocol LoginNetworkClientProtocol {
-
-    func login(requestBody: LoginRequestBodyNetworkModel) async throws -> LoginResponseNetworkModel
-
-}
-
-class LoginNetworkClient {
+class LoginNetworkClient: LoginNetworkClientProtocol {
 
     private let baseNetworkClient: BaseNetworkClientProtocol
 
     init(baseNetworkClient: BaseNetworkClientProtocol) {
         self.baseNetworkClient = baseNetworkClient
     }
-
-}
-
-extension LoginNetworkClient: LoginNetworkClientProtocol {
 
     func login(requestBody: LoginRequestBodyNetworkModel) async throws -> LoginResponseNetworkModel {
         try await baseNetworkClient.post(
