+++ QuizApp/Client/Login/LoginNetworkClientProtocol.swift
@@ -1,0 +1,5 @@
+protocol LoginNetworkClientProtocol {
+
+    func login(requestBody: LoginRequestBodyNetworkModel) async throws -> LoginResponseNetworkModel
+
+}
