+++ QuizApp/Data/Login/Repository/LoginRepositoryProtocol.swift
@@ -1,0 +1,5 @@
+protocol LoginRepositoryProtocol {
+
+    func login(data: LoginRequestBodyRepoModel) async throws -> LoginResponseRepoModel
+
+}
