+++ QuizApp/Domain/Login/LoginUseCase.swift
@@ -1,11 +1,5 @@
 import Foundation
 
-protocol LoginUseCaseProtocol {
-
-    func login(email: String, password: String) async throws
-
-}
-
 class LoginUseCase: LoginUseCaseProtocol {
 
     private let loginRepository: LoginRepositoryProtocol
