+++ QuizApp/Domain/Login/LoginUseCaseProtocol.swift
@@ -1,0 +1,5 @@
+protocol LoginUseCaseProtocol {
+
+    func login(email: String, password: String) async throws
+
+}
